requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
motor==3.3.1
pydantic[email]==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
